Updated headers

+++ include/helpers.h
@@ -1,7 +1,7 @@
 #ifndef HELPERS_H
 #define HELPERS_H
 
-char *input(char *msg, unsigned long long size);
+char *input(char *msg, size_t size);
 
 void clear_console();
